image-search-master/requirements.txt
Pillow
h5py
tensorflow
Keras
Flask
gevent
